requirements.txt
# Core scientific computing (fast execution)
numpy>=1.24.0
numba>=0.58.0

# Data handling
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress tracking
tqdm>=4.65.0
rich>=13.0.0

# Configuration management
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0.0

# Logging
structlog>=23.0.0

# CLI
typer>=0.9.0

# Testing
pytest>=7.4.0
pytest-benchmark>=4.0.0

# Serialization
orjson>=3.9.0

# Performance profiling
memory-profiler>=0.61.0
